fix(ui): streamline git changes refresh and interactions
Move git-changes state into a dedicated controller, refresh from shared session events, and simplify the git file list interactions with always-visible actions, full-path display, and prompt range insertion. Also improve API error parsing so SCM failures surface as clean user-facing messages.
